requirements.txt
paho-mqtt
service-bootstrap
homie-helpers~=0.0.6
